requirements.txt
python-telegram-bot[ext]
yt-dlp
python-dotenv
aiosqlite
